fix random4word, test addRemoveFav

+++ Project_CS163/dictionary.hpp
@@ -17,6 +17,7 @@ private:
     std::vector< TreeNode* > favorite_list;              // contain pointer to the end of word on the tree
     std::vector< TreeNode* > history;
     std::string pathCurrentDataset;
+  
     char def_delim;
     
 public:
@@ -25,12 +26,14 @@ public:
     Dictionary(std::string file_name, char delim)
     {
         pathCurrentDataset = file_name + ".txt";
-
-        if (!std::filesystem::exists(pathCurrentDataset))
+      
+        if (!std::filesystem::exists(pathCurrentDataset) 
+            && std::filesystem::exists("ORG_"+pathCurrentDataset))
             std::filesystem::copy_file("ORG_" + pathCurrentDataset, pathCurrentDataset);
 
         def_delim = delim;
         load(pathCurrentDataset, delim);
+        print_dic();
     }
     
     void load(std::string path, char delim_char) {
@@ -40,13 +43,13 @@ public:
         {
             std::string line;
             while (getline(fin, line)) {
+                
                 size_t delim = line.find(delim_char);
 
                 std::string word = line.substr(0, delim);
                 std::string def = line.substr(delim + 1, line.length() - delim - 1);
-
                 data.insert(word, def);
-
+                
                 for (std::string& str : util::str::split(def))
                 {
                     table.add(str, nullptr);
@@ -66,7 +69,7 @@ public:
 
         load(pathCurrentDataset, def_delim);
     }
-
+    
     void cache()
     {
         std::ofstream out(pathCurrentDataset);
@@ -127,6 +130,50 @@ public:
         std::cout << "Favorite list: " << '\n';
         int cnt = 0;
         for( TreeNode* address : favorite_list ) std::cout << ++cnt << ' ' << get_word(address) << '\n';
+    }
+    std::vector<TreeNode*> random4Word()
+    {
+       
+        std::vector<TreeNode*> result;
+        srand((unsigned int)time(0));
+        std::vector<int> index;
+        //random 4 number
+        while (index.size() < 4)
+        {
+            int t = rand() % size ;
+            bool check = false;
+            for (int i=0; i<index.size();++i)
+                if (t == index[i])
+                {
+                    check = true;
+                    break;
+                }
+            if (!check) index.push_back(t);
+        }
+        //sort 4 number
+        for (int i=0; i<index.size();++i)
+            for (int j=i;j<index.size(); ++j)
+                if (index[i] > index[j])
+                {
+                    int temp = index[i];
+                    index[i] = index[j];
+                    index[j] = temp;
+                }
+        std::vector<TreeNode*> temp = data.getListOfeow();
+        for (int i = 0; i < 4; ++i)
+            result.push_back(temp[index[i]]);
+        /*for (int i = 0; i < result.size(); ++i)
+        {
+            std::cout << get_word(result[i]) << "\n";
+        }*/
+        return result;
+    }
+    std::vector<TreeNode*> get()
+    {
+       
+        //return random4Word();
+        
+       
     }
 };
